Clean up all errors

+++ src/Core/Scope.c
@@ -207,7 +207,7 @@ static FloatRingBufferHandle getChannelBuffer(IScopeHandle scope, uint32_t chann
 static bool allChannelsAreStopped(ScopeHandle self){
     bool channelIsRunning = false;
 
-    for(int i = 0; i < self->amountOfChannels; ++i){
+    for(size_t i = 0; i < self->amountOfChannels; ++i){
         channelIsRunning = Channel_isRunning(self->channels[i]);
         if(channelIsRunning == true){
             break;
